requirements.txt
Django>=4.2,<4.3
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.3
